+++ src/3rdparty/opennurbs/zlib/zutil.c
@@ -275,7 +275,7 @@ void  zcfree (voidpf opaque, voidpf ptr)
 voidpf zcalloc (voidpf opaque, unsigned items, unsigned size)
 {
     if (opaque) opaque = 0; /* to make compiler happy */
-    return _halloc((long)items, size);
+    return _halloc((int)items, size);
 }
 
 void  zcfree (voidpf opaque, voidpf ptr)
